# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "kinoage",
        "description": "Documentation for kinoage admin dashboard and main site",
        "contact": {
            "email": "ayubxonnematov@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/studio": {
            "get": {
                "tags": [
                    "Studio"
                ],
                "summary": "Get list of Studio",
                "description": "Returns list of Studio",
                "operationId": "indexStudio",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Studio"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Studio"
                ],
                "summary": "Insert a new Studio",
                "description": "Insert a new Studio",
                "operationId": "storeStudio",
                "requestBody": {
                    "description": "Studio to create",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "$ref": "#/components/schemas/Studio"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Studio created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Studio"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation exception"
                    }
                }
            }
        },
        "/api/studio/{studio_id}": {
            "get": {
                "tags": [
                    "Studio"
                ],
                "summary": "Show a Studio from his Id",
                "description": "Show a Studio from his Id",
                "operationId": "showStudio",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Studio--id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Studio"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Studio not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Studio"
                ],
                "summary": "Delete a Studio",
                "description": "Delete a Studio",
                "operationId": "destroyStudio",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Studio--id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Studio"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Studio not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Studio"
                ],
                "summary": "Update an existing Studio",
                "description": "Update an existing Studio",
                "operationId": "updateStudio",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Studio--id"
                    }
                ],
                "requestBody": {
                    "description": "Studio to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "$ref": "#/components/schemas/Studio"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Studio"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Studio not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Studio": {
                "title": "Studio",
                "description": "Studio model",
                "required": [],
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id of Studio",
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "title": "name",
                        "description": "name of Studio",
                        "type": "string",
                        "example": "Macbook Pro"
                    },
                    "sku": {
                        "title": "sku",
                        "description": "sku of Studio",
                        "type": "string",
                        "example": "MCBPRO2022"
                    },
                    "price": {
                        "title": "price",
                        "description": "price of Studio",
                        "type": "integer",
                        "example": "99"
                    },
                    "created_at": {
                        "title": "created_at",
                        "type": "dateTime",
                        "readOnly": "true",
                        "example": "2022-07-04T02:41:42.336Z"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "type": "dateTime",
                        "readOnly": "true",
                        "example": "2022-07-04T02:41:42.336Z"
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
...
            "Studio--id": {
                "name": "Studio_id",
                "in": "path",
                "description": "Id of Studio",
                "required": true,
                "schema": {
                    "type": "integer",
                    "example": "1"
                }
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}